fix: edit about page invalid upload

+++ personalBlog/controllers/aboutController.py
@@ -28,19 +28,22 @@ def editAboutPostDef():
         name = form.name.data
         introduction = form.about_me.data
         email = form.email.data
+        filename = secure_filename(form.picture.data.filename)
+        if name is None or introduction is None or email is None or filename is None:
+            flash('Please fill in all fields', 'error')
+            return redirect(url_for('about.editAboutGet', form = form, message = '請填寫所有欄位'))
         # Check if a file was uploaded
         if form.picture.data:
             # Check if the file is a picture
             if not form.picture.data.filename.lower().endswith(('.png', '.jpg', '.jpeg')):
                 flash('File must be a picture', 'error')
-                return redirect(request.url)
+                return redirect('about/editAbout.html', form = form, message = '檔案不符合')
             
             # Save the picture
             if not os.path.exists(UPLOAD_FOLDER):
                 os.makedirs(UPLOAD_FOLDER)
             for filename in os.listdir(UPLOAD_FOLDER):
                 os.remove(os.path.join(UPLOAD_FOLDER, filename))
-            filename = secure_filename(form.picture.data.filename)
             form.picture.data.save(os.path.join( UPLOAD_FOLDER, filename))
             # Update the user's picture
         # Save the changes to the database
@@ -53,7 +56,8 @@ def editAboutPostDef():
             # The About table is not empty, so update the existing row
             about.name = name
             about.introduction = introduction
-            about.filename = filename
+            if filename is not None:
+                about.filename = filename
             about.email = email
             db.session.commit()
         db.session.commit()
@@ -63,5 +67,5 @@ def editAboutPostDef():
 
 
     # If the form data is not valid, redirect back to the editAbout page
-    flash('There was an error updating your profile', 'error')
+    flash('欄位不能為空，相片或著格式不符', 'error')
     return redirect(url_for('about.editAboutGet'))
